move plugin code into a single folder
