+++ application/configs/application.ini
@@ -1,4 +1,4 @@
-; ==============================================================================
+	; ==============================================================================
 [fundamental]
 ; ==============================================================================
 
@@ -96,8 +96,8 @@ doctrine.orm.entityManagers.default.metadataDrivers.drivers.0.annotationReaderCl
 doctrine.orm.entityManagers.default.metadataDrivers.drivers.0.annotationReaderCache = default
 
 doctrine.orm.entityManagers.default.metadataDrivers.drivers.1.adapterClass          = "Doctrine\ORM\Mapping\Driver\AnnotationDriver"
-doctrine.orm.entityManagers.default.metadataDrivers.drivers.1.mappingNamespace      = "CoreApi\Plugin"
-doctrine.orm.entityManagers.default.metadataDrivers.drivers.1.mappingDirs[]         = APPLICATION_PATH "/CoreApi/Plugin"
+doctrine.orm.entityManagers.default.metadataDrivers.drivers.1.mappingNamespace      = "Plugin"
+doctrine.orm.entityManagers.default.metadataDrivers.drivers.1.mappingDirs[]         = APPLICATION_PATH "/Plugin"
 doctrine.orm.entityManagers.default.metadataDrivers.drivers.1.annotationReaderClass = "Doctrine\Common\Annotations\SimpleAnnotationReader"
 doctrine.orm.entityManagers.default.metadataDrivers.drivers.1.annotationReaderCache = default
